+++ .gitignore
@@ -44,4 +44,3 @@ testem.log
 # System Files
 .DS_Store
 Thumbs.db
-environment.ts
