--- src/app/auth.service.ts
@@ -1,17 +1,0 @@
-import { Injectable } from '@angular/core';
-import * as firebase from 'firebase';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class AuthService {
-  constructor() { }
-  doRegister(value){
-    return new Promise<any>((resolve, reject) => {
-      firebase.auth().createUserWithEmailAndPassword(value.email, value.password)
-        .then(res => {
-          resolve(res);
-        }, err => reject(err));
-    });
-  }
-}
