+++ src/app/services/auth.service.ts
@@ -1,0 +1,27 @@
+import { Injectable } from '@angular/core';
+import * as firebase from 'firebase';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthService {
+  constructor() { }
+
+  doRegister(email: string, password: string){
+    return new Promise<any>((resolve, reject) => {
+      firebase.auth().createUserWithEmailAndPassword(email, password)
+        .then(res => {
+          resolve(res);
+        }, err => reject(err));
+    });
+  }
+
+  doSignIn(email: string, password: string){
+    return new Promise<any>((resolve, reject) => {
+      firebase.auth().signInWithEmailAndPassword(email, password)
+        .then(res => {
+          resolve(res);
+        }, err => reject(err));
+    });
+  }
+}
